+++ CHANGELOG.md
@@ -8,6 +8,53 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Released]
 
+## [0.2.0] - 2020-11-19
+### Changed
+- Bump version to 0.2.0
+- Update homepage
+
+## [0.1.9] - 2020-11-18
+### Changed
+- Art Component (A/B)
+
+## [0.1.8] - 2020-11-16
+### Changed
+- Art Component ()
+
+## [0.1.7] - 2020-11-13
+### Added
+- Art Component (placeholder for scripties)
+
+## [0.1.6] - 2020-11-03
+### Changed
+- Update header: tilde club 
+
+## [0.1.5] - 2020-10-27
+### Changed
+- Update homepage: tilde club 
+### Added
+- Size plugin json files
+
+## [0.1.4] - 2020-10-20
+### Changed
+- Update homepage
+
+## [0.1.3] - 2020-09-29
+### Added
+- Todo-list
+
+## [0.1.2] - 2020-09-28
+### Changed
+- Upgrade npm packages 
+- ncu -u
+- npm audit fix
+
+## [0.1.1] - 2020-09-25
+### Changed
+- Add profile Akiko
+
 ## [0.1.0] - 2020-06-24
 ### Added
 - This CHANGELOG file
+### Changed
+- Project name and version